+++ statemachine/statemachine.py
@@ -52,16 +52,20 @@ class CallableInstance(object):
 
 
 class Transition(object):
+    """
+    A transition holds reference to the source and destinations states.
+    """
 
-    def __init__(self, source, destination, identifier=None, validators=None):
-        self.source = source
-        self.destination = destination
-        self.identifier = identifier
-        self.validators = validators or []
+    def __init__(self, *states, **options):
+        self.source = states[0]
+        self.destinations = states[1:]
+        self.identifier = options.get('identifier')
+        self.validators = options.get('validators', [])
+        self.on_execute = options.get('on_execute')
 
     def __repr__(self):
         return "{}({!r}, {!r}, identifier={!r})".format(
-            type(self).__name__, self.source, self.destination, self.identifier)
+            type(self).__name__, self.source, self.destinations, self.identifier)
 
     def __or__(self, other):
         return CombinedTransition(self, other, identifier=self.identifier)
@@ -79,10 +83,16 @@ class Transition(object):
     def __set__(self, instance, value):
         "does nothing (not allow overriding)"
 
+    def __call__(self, f):
+        if not callable(f):
+            raise ValueError('Transitions can only be called as method decorators.')
+        self.on_execute = f
+        return self
+
     def _can_run(self, instance):
         return instance.current_state == self.source
 
-    def _run(self, instance, *args, **kwargs):
+    def _verify_can_run(self, instance):
         if not self._can_run(instance):
             raise LookupError(
                 _("Can't {} when in {}.").format(
@@ -91,12 +101,51 @@ class Transition(object):
                 )
             )
 
+    def _run(self, instance, *args, **kwargs):
+        self._verify_can_run(instance)
         self._validate(*args, **kwargs)
         return instance._activate(self, *args, **kwargs)
 
     def _validate(self, *args, **kwargs):
         for validator in self.validators:
             validator(*args, **kwargs)
+
+    def _get_destination(self, result):
+        """
+        A transition can point to one or more destination states.
+        If there is more than 1 state, the transition **must** specify a `on_execution` callback
+        that should return the desired state.
+        """
+        msg_multiple = _('Multiple destinations, you must return the desired state as: '
+                         '`return <State>` or `return <result>, <State>`.')
+
+        if len(self.destinations) == 1:
+            return result, self.destinations[0]
+
+        if result is None:
+            raise ValueError(msg_multiple)
+        elif isinstance(result, State):
+            destination = result
+            result = None
+        else:
+            try:
+                num = len(result)
+            except TypeError:
+                raise ValueError(msg_multiple)
+
+            if num != 2:
+                raise ValueError(msg_multiple)
+
+            result, destination = result
+
+        if not isinstance(destination, State):
+            raise ValueError(msg_multiple)
+
+        if destination not in self.destinations:
+            raise ValueError(
+                _('{} is not a possible destination state.').format(destination.name))
+
+        return result, destination
 
 
 class CombinedTransition(Transition):
@@ -107,7 +156,13 @@ class CombinedTransition(Transition):
 
     @property
     def _right(self):
-        return self.destination
+        return self.destinations[0]
+
+    def __call__(self, f):
+        result = super(CombinedTransition, self).__call__(f)
+        self._left.on_execute = self.on_execute
+        self._right.on_execute = self.on_execute
+        return result
 
     def __contribute_to_class__(self, managed, identifier):
         super(CombinedTransition, self).__contribute_to_class__(managed, identifier)
@@ -118,14 +173,7 @@ class CombinedTransition(Transition):
         return instance.current_state in [self._left.source, self._right.source]
 
     def _run(self, instance, *args, **kwargs):
-        if not self._can_run(instance):
-            raise LookupError(
-                _("Can't {} when in {}.").format(
-                    self.identifier,
-                    instance.current_state.name
-                )
-            )
-
+        self._verify_can_run(instance)
         self._validate(*args, **kwargs)
         transition = self._left if instance.current_state == self._left.source else self._right
         return transition._run(instance, *args, **kwargs)
@@ -144,8 +192,8 @@ class State(object):
         return "{}({!r}, identifier={!r}, value={!r}, initial={!r})".format(
             type(self).__name__, self.name, self.identifier, self.value, self.initial)
 
-    def to(self, state):
-        transition = Transition(self, state)
+    def to(self, *states):
+        transition = Transition(self, *states)
         self.transitions.append(transition)
         return transition
 
@@ -256,9 +304,12 @@ class BaseStateMachine(object):
         self.current_state_value = value.value
 
     def _activate(self, transition, *args, **kwargs):
-        on_event = getattr(self, 'on_{}'.format(transition.identifier), None)
-        result = on_event(*args, **kwargs) if callable(on_event) else None
-        self.current_state = transition.destination
+        on_event = getattr(self, 'on_{}'.format(transition.identifier), transition.on_execute)
+        result = on_event(self, *args, **kwargs) if callable(on_event) else None
+
+        result, destination = transition._get_destination(result)
+
+        self.current_state = destination
         return result
 
     def get_transition(self, transition_identifier):