+++ HISTORY.rst
@@ -7,6 +7,8 @@ History
 
 * Python 3.6 support.
 * Drop official support for Python 3.3.
+* `Transition` can be used as decorator for `on_execute` callback definition.
+* `Transition` can point to multiple destination states.
 
 
 0.3.0 (2017-03-22)